.agent/requirements.txt
# Personal AI OS - Master Requirements
# Install: pip install -r .agent/requirements.txt

# Core
pyyaml>=6.0
requests>=2.31.0
python-frontmatter>=1.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Git operations (optional)
gitpython>=3.1.0

# Dev.to API
# (uses requests)

# Linear API
# (uses requests)

# Obsidian Local REST API
# (uses requests)
